Create global list page and prepare FE for admin access

+++ scavengerHunt/src/components/ItemsList.js
@@ -67,40 +67,54 @@ class ItemsList extends Component {
     this.props.deleteItem(item, this.props.list);
   }
 
+  // Shows the list of items on the page
   renderList() {
-    const { nameStyle, descriptionStyle } = styles;
-
     if (this.props.list.items) {
       return this.props.list.items.map((item, index) => {
-        const swipeButtons = [{
-          key: Math.random(),
-          text: 'Delete',
-          backgroundColor: 'red',
-          onPress: () => this.deleteItem(item)
-        }];
+        if (this.props.list.admin) {
+          const swipeButtons = [{
+            key: Math.random(),
+            text: 'Delete',
+            backgroundColor: 'red',
+            onPress: () => this.deleteItem(item)
+          }];
+          return (
+            <Swipeout key={index} right={swipeButtons}>
+              {this.renderBody(item, index)}
+              <Button onPress={this.addItemToList.bind(this)}>Add Item</Button>
+            </Swipeout>
+          );
+        }
+
         return (
-          <Swipeout key={index} right={swipeButtons}>
-          <CardSection style={{ borderBottomWidth: 0, padding: 20, height: 100 }}>
-            { this.isComplete(item.complete, item) }
-            <Text
-              style={nameStyle}
-              onPress={() => {
-                if (item.complete === 0) {
-                  this.uncheckedBoxClicked(item);
-                }
-              }}
-            >{ `${item.name} ${'\n'}` }
-            <Text
-              style={descriptionStyle}
-            >{`${item.description}`}
-            </Text>
-            </Text>
-          </CardSection>
-          </Swipeout>
+          this.renderBody(item, index)
         );
       });
     }
     return;
+  }
+
+  // Allows editing based on whether the current user made the list
+  renderBody(item, index) {
+    const { nameStyle, descriptionStyle } = styles;
+    return (
+      <CardSection key={index} style={{ borderBottomWidth: 0, padding: 20, height: 100 }}>
+        { this.isComplete(item.complete, item) }
+        <Text
+          style={nameStyle}
+          onPress={() => {
+            if (item.complete === 0) {
+              this.uncheckedBoxClicked(item);
+            }
+          }}
+        >{ `${item.name} ${'\n'}` }
+        <Text
+          style={descriptionStyle}
+        >{`${item.description}`}
+        </Text>
+        </Text>
+      </CardSection>
+    );
   }
 
   render() {
@@ -108,9 +122,8 @@ class ItemsList extends Component {
       <ScrollView style={styles.listStyle}>
         <Card>
           { this.renderList() }
-      <Button onPress={this.addItemToList.bind(this)}>Add Item</Button>
-      </Card>
-    </ScrollView>
+        </Card>
+      </ScrollView>
     );
   }
 }
